set namespace
