+++ CollectSFData/Common/CustomTaskManager.cs
@@ -9,7 +9,7 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace CollectSFData
+namespace CollectSFData.Common
 {
     public class CustomTaskManager : Instance
     {
