+++ CollectSFData/DataFile/DtrTraceRecord.cs
@@ -5,18 +5,18 @@
 
 using System;
 
-namespace CollectSFData
+namespace CollectSFData.DataFile
 {
     [Serializable]
-    public class CsvSetupRecord : ITraceRecord
+    public class DtrTraceRecord : ITraceRecord
     {
-        private const int _fieldCount = 5;
+        private const int _fieldCount = 6;
 
-        public CsvSetupRecord()
+        public DtrTraceRecord()
         {
         }
 
-        public CsvSetupRecord(string traceRecord, FileObject fileObject, string resourceUri = null)
+        public DtrTraceRecord(string traceRecord, FileObject fileObject, string resourceUri = null)
         {
             Populate(fileObject, traceRecord, resourceUri);
         }
@@ -41,15 +41,16 @@ namespace CollectSFData
 
         public string Type { get; set; }
 
-        public IRecord Populate(FileObject fileObject, string traceRecord, string resourceUri = null)
+        public IRecord Populate(FileObject fileObject, string dtrRecord, string resourceUri = null)
         {
-            string[] fields = ParseRecord(traceRecord);
+            string[] fields = ParseRecord(dtrRecord);
 
-            Timestamp = Convert.ToDateTime(fields[0].Replace("-", " "));
+            Timestamp = Convert.ToDateTime(fields[0]);
             Level = fields[1];
-            PID = Convert.ToInt32(fields[2]);
-            Type = fields[3];
-            Text = fields[4];
+            TID = Convert.ToInt32(fields[2]);
+            PID = Convert.ToInt32(fields[3]);
+            Type = fields[4];
+            Text = fields[5];
             NodeName = fileObject.NodeName;
             FileType = fileObject.FileDataType.ToString();
             RelativeUri = fileObject.RelativeUri;
@@ -60,24 +61,26 @@ namespace CollectSFData
 
         public override string ToString()
         {
-            return $"{Timestamp:o},{Level},{PID},{Type},{Text},{NodeName},{FileType},{RelativeUri},{ResourceUri}{Environment.NewLine}";
+            return $"{Timestamp:o},{Level},{TID},{PID},{Type},{Text},{NodeName},{FileType},{RelativeUri},{ResourceUri}{Environment.NewLine}";
         }
 
         private string[] ParseRecord(string record)
         {
             // format for csv compliance
-            // kusto conforms to csv standards. service fabric trace (csv file) does not
+            // by default the Text field is not quoted, contains commas, contains quotes
+            // kusto conforms to csv standards. service fabric dtr.zip (csv file) does not
 
             string[] newLine = record.Split(new string[] { "," }, _fieldCount, StringSplitOptions.None);
             string additionalCommas = string.Empty;
+            string newText = newLine[newLine.Length - 1];
 
             if (newLine.Length < _fieldCount)
             {
                 additionalCommas = new string(',', _fieldCount - newLine.Length);
             }
 
-            newLine[newLine.Length - 1] = newLine[newLine.Length - 1].Replace("\"", "'").TrimEnd('\r', '\n');
-            newLine[newLine.Length - 1] = $"{additionalCommas}\"{newLine[newLine.Length - 1]}\"";
+            newText = newText.TrimStart('\"').TrimEnd('\"', '\r', '\n').Replace("\"", "'");
+            newLine[newLine.Length - 1] = $"{additionalCommas}\"{newText}\"";
             return newLine;
         }
     }
