+++ CollectSFData/LogAnalytics/LogAnalyticsPurge.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
 // ------------------------------------------------------------
 
-namespace CollectSFData
+namespace CollectSFData.LogAnalytics
 {
     public class LogAnalyticsPurge
     {
